pubsub_app/pyproject.toml
[tool.poetry]
name = "pubsub-app"
version = "0.1.0"
description = ""
authors = ["Erasmo <erasmodimarco@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-pubsub = "^2.21.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-cov = "^5.0.0"

[tool.poetry.scripts]
test = "pytest"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
